refactor(preload): modularize preload/index.ts into apis/ modules
Split the 965-line preload API object into 13 focused modules under
src/preload/apis/ mirroring the main ipc module structure:
- app, system, file, fileUtils, backup, knowledge, auth, agents,
  integrations, notification, utility, window, data, and shared.

The api object is composed via spread in index.ts, preserving
WindowApiType inference and contextBridge behavior.
