+++ src/preload/apis/shared.ts
@@ -1,0 +1,10 @@
+import type { SpanContext } from '@opentelemetry/api'
+import { ipcRenderer } from 'electron'
+
+export function tracedInvoke(channel: string, spanContext: SpanContext | undefined, ...args: any[]) {
+  if (spanContext) {
+    const data = { type: 'trace', context: spanContext }
+    return ipcRenderer.invoke(channel, ...args, data)
+  }
+  return ipcRenderer.invoke(channel, ...args)
+}
