+++ src/preload/apis/utility.ts
@@ -1,0 +1,15 @@
+import { IpcChannel } from '@shared/IpcChannel'
+import { ipcRenderer } from 'electron'
+
+export const utilityApi = {
+  zip: {
+    compress: (text: string) => ipcRenderer.invoke(IpcChannel.Zip_Compress, text),
+    decompress: (text: Buffer) => ipcRenderer.invoke(IpcChannel.Zip_Decompress, text)
+  },
+  aes: {
+    encrypt: (text: string, secretKey: string, iv: string) =>
+      ipcRenderer.invoke(IpcChannel.Aes_Encrypt, text, secretKey, iv),
+    decrypt: (encryptedData: string, iv: string, secretKey: string) =>
+      ipcRenderer.invoke(IpcChannel.Aes_Decrypt, encryptedData, iv, secretKey)
+  }
+}
